Build PHP 7.2 with libsodium by default
I don't understand the structure of brew formulae completely, but it seems that when the `depends_on "libsodium" => :recommended if name.split("::")[2].downcase.start_with?("php72")` is added, homebrew will automatically add an option `--without-libsodium`. With that said, the current check `if build.with? "sodium"` never gets to true. Changing to `unless build.without? "libsodium"` works for me locally.